Finished Creating Homepage and Contact Us
Finished creating the homepage and contact us page. All necessary css
files for this have been updated. And necessary components have been
modified and added as needed. The context files have also been created
and filled with the appropriate content. Global context has also been
defined for the navbar but it has yet to be linked.