Fixed Social Banner & Routes & Added Icons
The social banner is now properly integrated in the globals file. There
is also the routes which has been moved to its own templates file along
with the social templates. This means the individual page files no
longer needed their own social message because it's in globals.
The appropriate social icons [ Twitter, Instagram, Facebook, Google+]
are now also functioning in the social banner.